+++ UIKit/Classes/UITouch.m
@@ -148,7 +148,7 @@
             phase = @"Right-clicked";
             break;
     }
-    return [NSString stringWithFormat:@"<%@: %p; timestamp = %e; tapCount = %d; phase = %@; view = %@; window = %@>", [self className], self, self.timestamp, self.tapCount, phase, self.view, self.window];
+    return [NSString stringWithFormat:@"<%@: %p; timestamp = %e; tapCount = %d; phase = %@; view = %@; window = %@>", [self className], self, self.timestamp, (int)self.tapCount, phase, self.view, self.window];
 }
 
 @end
